Wire provider config through RAG modules
